refactor: introduced automatic loading of icons
In the original repo as well as your forked repo,
it would take the programmer a number of steps in
order to add a skill icon (check my resume at
https://orwa.tech). Namely, the programmer would
have to add an icon file in the src/icons folder,
import it into the Skills section, then add it to
the lookup dictionary, there is also the issue of
lowercase vs. uppercase, and the number of other
places where we also need a unified way of using
different icons (such as in the Hero section).

As a unified and an elegant solution, I am propos
-ing the use of `import.meta.blob` to import all
the files ending with `.astro` inside of the icons
folder, and then, using a Proxy (an interceptor of
array [] lookup) return the AstroInstance in a case
-insensitive manner.

I have used it in my site, and I wondered why I did
not do it earlier, since the code was growing un
-controllably as I was adding more and more icons.

Note that the code I am requesting you to pull is
much cleaner than the one I have on my site. I app
-reciate the amount of work you have put into
bringing i18n to this great template and hence,
consider this pull request a token of gratitude
for your work.

Regards,
Orwa.